[frontend] Nested ref creation on observable knowledge working
